A few bug fixes for christmas cards (#428)
* Switch joel to using active row

* Fix biffa not working with damage effects

* Fix problems with swapRows function

* Fix problems with swapRows function

* Fix trapdoor hooks

* Switch lantern/glowstone to popping up again when X-ed out of
